app/requirements.txt
httpx
telethon
python-dotenv
apscheduler
tweepy